Switches to flask-sqlalchemy
This should fix some issues with the API crashing all the times with the
"object fetched on a different thread than created it" issues.
