+++ bannedWordServer/models/__init__.py
@@ -1,3 +1,2 @@
-from sqlalchemy.ext.declarative import declarative_base
-
-Base = declarative_base()
+from .ban import Ban
+from .server import Server
