+++ bannedWordServer/models/ban.py
@@ -1,16 +1,15 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
 from datetime import datetime, timedelta
 
-from bannedWordServer.models import Base
+from bannedWordServer import db
 
-class Ban(Base):
+class Ban(db.Model):
 	__tablename__ = 'server_banned_word'
 
-	rowid = Column(Integer, primary_key=True)
-	banned_word = Column(String, nullable=False, default='defaultbannedword')
-	server_id = Column(Integer, ForeignKey('server.server_id'), nullable=False)
-	infracted_at = Column(String, nullable=False, default=datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
-	calledout_at = Column(String, nullable=False, default=(datetime.now() - timedelta(weeks=52)).strftime("%Y-%m-%d %H:%M:%S"))
+	rowid = db.Column(db.Integer, primary_key=True)
+	banned_word = db.Column(db.String, nullable=False, default='defaultbannedword')
+	server_id = db.Column(db.Integer, db.ForeignKey('server.server_id'), nullable=False)
+	infracted_at = db.Column(db.String, nullable=False, default=datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
+	calledout_at = db.Column(db.String, nullable=False, default=(datetime.now() - timedelta(weeks=52)).strftime("%Y-%m-%d %H:%M:%S"))
 
 	def to_dict(self):
 		entries = {}
