+++ bannedWordServer/routes/__init__.py
@@ -1,0 +1,3 @@
+from .banroute import BanRoute
+from .messageroute import MessageRoute
+from .serverroute import ServerRoute
