+++ requirements.txt
@@ -1,6 +1,7 @@
 Click==7.0
 Flask==1.1.1
 Flask-Cors==3.0.8
+Flask-SQLAlchemy==2.4.1
 itsdangerous==1.1.0
 Jinja2==2.10.1
 MarkupSafe==1.1.1
