+++ tests/routes/test_serverroute.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 
 from bannedWordServer.config import BOT_TOKEN
 from bannedWordServer.constants.errors import NotFoundError, InvalidTypeError, DuplicateResourceError, AuthenticationError
-from bannedWordServer.models import Base
+from bannedWordServer import db
 from bannedWordServer.models.server import Server
 from bannedWordServer.models.ban import Ban
 from bannedWordServer.routes.serverroute import ServerRoute
@@ -14,7 +14,7 @@ engine = create_engine('sqlite:///:memory:')
 
 class TestServerRouteGetCollection(TestCase):
 	def setUp(self):
-		Base.metadata.create_all(engine)
+		db.Model.metadata.create_all(engine)
 		self.connection = engine.connect()
 		self.trans = self.connection.begin()
 		self.session = Session(bind=self.connection)
@@ -53,7 +53,7 @@ class TestServerRouteGetCollection(TestCase):
 
 class TestServerRouteGetOne(TestCase):
 	def setUp(self):
-		Base.metadata.create_all(engine)
+		db.Model.metadata.create_all(engine)
 		self.connection = engine.connect()
 		self.trans = self.connection.begin()
 		self.session = Session(bind=self.connection)
@@ -94,7 +94,7 @@ class TestServerRouteGetOne(TestCase):
 
 class TestServerRoutePostCollection(TestCase):
 	def setUp(self):
-		Base.metadata.create_all(engine)
+		db.Model.metadata.create_all(engine)
 		self.connection = engine.connect()
 		self.trans = self.connection.begin()
 		self.session = Session(bind=self.connection)
@@ -129,7 +129,7 @@ class TestServerRoutePostCollection(TestCase):
 
 class TestServerRoutePartialUpdate(TestCase):
 	def setUp(self):
-		Base.metadata.create_all(engine)
+		db.Model.metadata.create_all(engine)
 		self.connection = engine.connect()
 		self.trans = self.connection.begin()
 		self.session = Session(bind=self.connection)
